[feat] split get image functions to component
